+++ src/main/java/dev/willyelton/crystal_tools/client/gui/UpgradeScreen.java
@@ -61,7 +61,7 @@ public class UpgradeScreen extends BaseUpgradeScreen {
     protected void resetPoints(boolean crystalRequired) {
         if (!crystalRequired || this.player.getInventory().hasAnyOf(Set.of(Registration.CRYSTAL.get()))) {
             PacketDistributor.sendToServer(ResetSkillsPayload.INSTANCE);
-            ToolUtils.resetPoints(this.stack);
+            ToolUtils.resetPoints(this.stack, this.player);
 
             data = ToolUtils.getSkillData(this.stack);
         }